[BE] refactor: group api 리팩토링
